requirements.txt
grpcio==1.55.0
grpcio-tools==1.55.0
protobuf==4.23.1